First Revision
I changed the title because I think it's more correct to say "What are the differences..." when there are more than 2 things you're comparing.

My first question for the author would be, "Who is the audience?" I would ask this so I could better understand how to construct the writing. It would allow me to understand how much my audience might know about the topic. This would help me know if I needed to include more basic information to help their understanding.

I noticed a lot of inaccuracies in the original content. In fact, I believe about 6 sentences in the original content are not true. This leads me to believe that the original author is not an SME themselves. In cases like this, I would then take more liberty to rewrite the original content. On the contrary, if the original author was more of an expert, I would take more of an editing role on the content.

My second question for the original author would be, "What do you want your reader to be able to do after they read this?" This is important because it could change this writing from a conceptual piece to a task. For example, if the author told me "I want the reader to be able to fetch changes, merge them, make their own changes, then push them back to the remote repo..." - Then I would instead format this as a task with numbered steps.