Second Revision
In this revision, I was tying to think of a way to make the reader understand that `git push` is very different from the other two commands. And I wanted to make the reader understand that pull and fetch are very similar. For that reason, I added two introductory sentences to try and explain this.

I also decided to start breaking the content into more sections using Markdown headers. I think breaking content into subsections is a great way to make writing more digestible, especially for those readers who are just scanning for information they need to find.

At this point I would ask the original author questions like "How important is it for the reader to know the difference between git fetch and git pull". The answer would inform me if it is even worth it  to have a subsection for this.

Overall, I'm not too happy with the direction I went down with this commit. I think I've convoluted the writing and made it too wordy. I liked my theory of adding more Markdown headers to break the content up, but it didn't work here. So I think in my next revision I'll use more Markdown headers, but take a more basic approach to formatting this information.