Third Revision
I removed the "differences between git fetch and git pull" section, and decided to add one section for each command. I think this makes it easier to understand the differences and scan for information; not to mention it more resembles the format of the author's original draft.

I also rearranged the order of the commands into the order a user might actually make these commands in the real world. However, I'm not sure this brings much value.