Fourth Revision
I removed the first two introductory sentences. I think they were redundant to have, and the content was more appropriately explained within the subsection for each command.

This also helped inform me to change the title. I realized that this is more of a comparison of all the commands, not just writing about the differences. This is especially important because I think it's important for the reader to understand the similarities of pull and fetch. And if we're talking about similarities and differences in a piece of writing, then it's more appropriately titled as a "comparison".

I added hyperlinks to terms I think the reader might need to understand the writing. I added hyperlinks to good sources for the terms "tracking branch" and "remote repository". I think if the reader doesn't understand the differences between the three commands we're discussing, then they'll probably need to learn what these terms mean too.

I rearranged the order of the git pull and git fetch sections, because the git pull section mentions git fetch and git merge. So I thought it should go after I explain those things so the reader knows what I'm talking about.

I changed the subtitle of the last section to "When to use each command" from "In summary:", because I think this is more descriptive of what the section explains. I debated removing this section completely, because it is slightly redundant. And redundancy in technical writing is usually highly frowned upon. However, I learned as a teacher that when you're trying to educate someone (which is what we're doing here), then it's okay to repeat important things, especially if you can say them again in slightly different ways. Repeating important concepts in slightly different ways can especially help folks with learning disabilities to grasp concepts.

I added a "Syntax:" bullet point to each section. I debated as to whether this was adding unnecessary content, or too advanced content. But I remembered in the original draft the author did include syntax for "git merge", so I thought it couldn't hurt, and might actually help someone practice using the commands.